% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_pairwise_distances_XCall}
\alias{cpp_pairwise_distances_XCall}
\title{cpp_pairwise_distances_XCall}
\usage{
cpp_pairwise_distances_XCall(X,metric)
}
\arguments{
\item{X}{[Rcpp::NumericMatrix] Matrix}

\item{metric}{[Rcpp::Function] R function}
}
\description{
Pairwise distances between X and themselves with a R function (metric).
DO NOT USE, use SBCK::pairwise_distances
}
