% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_pairwise_distances_Xstr}
\alias{cpp_pairwise_distances_Xstr}
\title{cpp_pairwise_distances_Xstr}
\usage{
cpp_pairwise_distances_Xstr(X,str_metric)
}
\arguments{
\item{X}{[Rcpp::NumericMatrix] Matrix}

\item{str_metric}{[std::string] c++ string}
}
\description{
Pairwise distances between X and themselves with a compiled 
str_metric. DO NOT USE, use SBCK::pairwise_distances
}
