% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_pairwise_distances_XYCall}
\alias{cpp_pairwise_distances_XYCall}
\title{cpp_pairwise_distances_XYCall}
\usage{
cpp_pairwise_distances_XYCall(X,Y,metric)
}
\arguments{
\item{X}{[Rcpp::NumericMatrix] Matrix}

\item{Y}{[Rcpp::NumericMatrix] Matrix}

\item{metric}{[Rcpp::Function] R function}
}
\description{
Pairwise distances between X  and Y with a R function (metric).
DO NOT USE, use SBCK::pairwise_distances
}
