% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_pairwise_distances_XYstr}
\alias{cpp_pairwise_distances_XYstr}
\title{cpp_pairwise_distances_XYstr}
\usage{
cpp_pairwise_distances_XYstr(X,Y,str_metric)
}
\arguments{
\item{X}{[Rcpp::NumericMatrix] Matrix}

\item{Y}{[Rcpp::NumericMatrix] Matrix}

\item{str_metric}{[std::string] c++ string}
}
\description{
Pairwise distances between two differents matrix X and Y with a 
compiled str_metric. DO NOT USE, use SBCK::pairwise_distances
}
