% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.univariates.R
\name{dataset_gaussian_exp_mixture_1d}
\alias{dataset_gaussian_exp_mixture_1d}
\title{dataset_gaussian_exp_mixture_1d}
\usage{
dataset_gaussian_exp_mixture_1d(n_samples)
}
\arguments{
\item{n_samples}{[integer] numbers of samples drawn}
}
\value{
[list] a list containing X0, X1 (biased in calibration/projection)
        and Y0 (reference in calibration)
}
\description{
Generate a univariate testing dataset from a mixture of gaussian and
exponential distribution
}
\examples{
XY = SBCK::dataset_gaussian_exp_mixture_1d(2000)
XY$X0 ## Biased in calibration period
XY$Y0 ## Reference in calibration period
XY$X1 ## Biased in projection period

}
