% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.multivariates.R
\name{dataset_gaussian_L_2d}
\alias{dataset_gaussian_L_2d}
\title{dataset_gaussian_L_2d}
\usage{
dataset_gaussian_L_2d(n_samples)
}
\arguments{
\item{n_samples}{[integer] numbers of samples drawn}
}
\value{
[list] a list containing X0, X1 (biased in calibration/projection)
        and Y0 (reference in calibration)
}
\description{
Generate a testing dataset such that the biased dataset is a normal
distribution and reference a mixture a normal with a form in "L"
}
\examples{
XY = SBCK::dataset_gaussian_L_2d(2000)
XY$X0 ## Biased in calibration period
XY$Y0 ## Reference in calibration period
XY$X1 ## Biased in projection period

}
