% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.shuffle.R
\name{schaake_shuffle}
\alias{schaake_shuffle}
\title{schaake_shuffle function}
\usage{
schaake_shuffle(Y0,X0)
}
\arguments{
\item{Y0}{[vector] The reference vector}

\item{X0}{[vector] The vector to transform the rank}
}
\value{
Z0 [vector] X shuffled.
}
\description{
Apply the Schaake shuffle to transform the rank of X0 such that its
correspond to the rank of Y0
}
\examples{
X0 = stats::runif(10)
Y0 = stats::runif(10)
Z0 = SBCK::schaake_shuffle( Y0 , X0 )

}
