% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.where.R
\name{where}
\alias{where}
\title{where function}
\usage{
where(cond,x,y)
}
\arguments{
\item{cond}{[vector/matrix/array] Boolean values}

\item{x}{[vector/matrix/array] Values if cond is TRUE}

\item{y}{[vector/matrix/array] Values if cond is FALSE}
}
\value{
z [vector/matrix/array].
}
\description{
This function return a vector / matrix / array of the same shape than
cond / x / y such that if(cond) values are x, and else y.
}
\examples{
x = base::seq( -2 , 2 , length = 100 )
y = where( x < 1 , x , exp(x) ) ## y = x if x < 1, else exp(x)

}
